A lot more added.
Mostly added from https://wiki.osdev.org/Meaty_Skeleton

This requires significantly more work for it to be considered at a
beginning release state.

It is nice to have the submodules in there for gcc and binutils, but
ultimatly you can create that stuff on your own.

Ideally, I'll create a script that will automatically build gcc and
binutils for a given architecture, but that's out of scope for right now.

I've yet to create a "true" 386 machine, but I may wind up moving this
to the m68k and using an the 6850 ACIA as the terminal I/O (tty.c).

Will need to put in some KiCAD stuff when that happens though. I'd like
to use the m68k20 or higher since it has virtual memory, but that is
propably outside of my skill set at this current point in time. This
leaves the m68000 or m68008 as SBCs for whatever I write. That's fine,
because I think gcc can target a vanilla m68000.

I don't think there is a C compiler for the 6502 so that means I
probably won't target that, because I lack the knowledge to create a
"true" C compiler for that architecture (let alone one that has all the
modern features in C11). However, that might be more of an issue of
working with binutils than gcc(?)